abstracts the create share link functionality
